move new homepage out of staging
